Fixes all around, mainly to minor annoyances
Added some things to the style sheet to prevent chrome from being weird, deleted calendar for now, it was just an empty page. Fixed formatting for various appointments pages.

!!! A BIG change that needs to be done for the database: !!!
Change the enum for the 'Status' of appointments with the following SQL:

ALTER TABLE appointments
MODIFY COLUMN Status ENUM('Pending', 'Approved', 'Rejected', 'Finished') NOT NULL DEFAULT 'Pending';

This is mainly for simplifying things since we don't handle billing and whatnot and need not care if it was cancelled or a no-show.